Authorization validation is handled with JsonWebToken with the help of ValidateTokenHnadler middleware. Contact route is needed authorization bearer key.

+++ controllers/contactController.js
@@ -5,15 +5,15 @@ import AppError from '../util/AppError.js';
 
 //@desc Get all contacts
 //@route GET /api/contacts
-//@access public
+//@access private
 export const getAllContacts = asyncHandler(async (req, res, next) => {
-    const contacts = await Contact.find();
+    const contacts = await Contact.find({ user_id: req.user.id });
     res.status(200).json(contacts);
 });
 
 //@desc Get a contact
 //@route GET /api/contacts/:id
-//@access public
+//@access private
 export const getContact = asyncHandler(async (req, res, next) => {
     const { id } = req.params;
     // Önce ID'nin geçerli olup olmadığını kontrol et
@@ -30,7 +30,7 @@ export const getContact = asyncHandler(async (req, res, next) => {
 
 //@desc Create new contact
 //@route POST /api/contacts
-//@access public
+//@access private
 export const postContact = asyncHandler(async (req, res, next) => {
     const { name, email, phone } = req.body;
 
@@ -41,13 +41,18 @@ export const postContact = asyncHandler(async (req, res, next) => {
     }
 
     // Contact.create() hem dokümanı oluşturur hem de veritabanına kaydeder
-    const contact = await Contact.create({ name, email, phone });
+    const contact = await Contact.create({
+        user_id: req.user.id,
+        name,
+        email,
+        phone,
+    });
     res.status(201).json(contact);
 });
 
 //@desc Update a contact
 //@route PUT /api/contacts/:id
-//@access public
+//@access private
 export const putContact = asyncHandler(async (req, res, next) => {
     const { id } = req.params;
     // Geçerli bir ID olup olmadığını kontrol et
@@ -55,21 +60,32 @@ export const putContact = asyncHandler(async (req, res, next) => {
         return next(new AppError('Invalid contact ID format', 400));
     }
 
-    const contact = await Contact.findByIdAndUpdate(id, req.body, {
-        new: true,
-        runValidators: true,
-    });
+    const contact = await Contact.findById(req.params.id);
 
     if (!contact) {
         return next(new AppError('Contact is not found.', 404));
     }
 
-    res.status(200).json(contact);
+    if (contact.user.user_id.toString() !== req.user.id) {
+        return next(
+            new AppError(
+                "User don't have permission to update other users contacts.",
+                403
+            )
+        );
+    }
+
+    const updatedContact = await Contact.findByIdAndUpdate(id, req.body, {
+        new: true,
+        runValidators: true,
+    });
+
+    res.status(200).json(updatedContact);
 });
 
 //@desc Delete a contact
 //@route DELETE /api/contacts/:id
-//@access public
+//@access private
 export const deleteContact = asyncHandler(async (req, res, next) => {
     const { id } = req.params;
     // Önce ID'nin geçerli olup olmadığını kontrol et
@@ -77,10 +93,32 @@ export const deleteContact = asyncHandler(async (req, res, next) => {
         return next(new AppError('Invalid contact ID format', 400));
     }
 
-    const contact = await Contact.findByIdAndDelete(id);
+    const contact = await Contact.findById(req.params.id);
+
     if (!contact) {
         return next(new AppError('Contact is not found.', 404));
     }
 
-    res.status(200).json({ message: 'Contact deleted successfully!', contact });
+    if (contact.user.user_id.toString() !== req.user.id) {
+        return next(
+            new AppError(
+                "User don't have permission to update other users contacts.",
+                403
+            )
+        );
+    }
+    const deletedContact = await Contact.findByIdAndDelete(id);
+    if (!deletedContact) {
+        return next(
+            new AppError(
+                "User don't have permission to delete other users contacts.",
+                404
+            )
+        );
+    }
+
+    res.status(200).json({
+        message: 'Contact deleted successfully!',
+        deletedContact,
+    });
 });
